LayoutTests/platform/gtk/fast/text/selection-in-initial-advance-region-expected.txt

This tests for a drag select crasher.

لاً
PASS sel.rangeCount is 1
PASS r.startContainer is r.endContainer
PASS r.startOffset is 0
FAIL r.endOffset should be 3. Was 1.
